vision-service/requirements.txt
requests
websocket-client
Pillow
fastapi
uvicorn